# ♻️ Refactor: Improve Performance of getSplicedStrList